requirements.txt
aniso8601==3.0.2
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coveralls==1.7.0
cssutils==1.0.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.8
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-facebook-login==4.0.0
django-filter==1.1.0
django-heroku==0.3.1
django-inlinecss==0.1.2
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-sendgrid-v5==0.7.1
djangorestframework==3.9.2
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.14.0
entrypoints==0.3
facebook-sdk==3.1.0
flake8==3.7.7
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
inflection==0.3.1
iso8601==0.1.12
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
more-itertools==6.0.0
nose==1.3.7
oauthlib==3.0.1
pbr==5.1.3
Pillow==5.4.1
pluggy==0.9.0
promise==2.2.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pynliner3==0.6
pytest==4.3.1
pytest-cov==2.6.1
python-decouple==3.1
python-http-client==3.1.0
python-twitter==3.5
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.15.89
Rx==1.6.1
sendgrid==5.6.0
singledispatch==3.4.0.3
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
soupsieve==1.9
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.3
whitenoise==4.1.2
wrapt==1.11.1
